pyproject.toml
[tool.poetry]
name = "async-fivem"
version = "1.0.0"
description = "Async API for FiveM endpoints"
authors = ["makubob <makupi@icloud.com>"]
license = "GPL-3.0"
packages = [{include = "fivem"}]

[tool.poetry.dependencies]
python = "^3.6"
aiohttp = "\b"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
aiohttp = "^3.6.2"
python-dotenv = "^0.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
